requirements.txt
# Core Libraries
# torch==<version_compatible_with_cuda> --extra-index-url https://download.pytorch.org/whl/<cuda_version_tag>
# torchaudio==<version>
torch==2.4.0 --extra-index-url https://download.pytorch.org/whl/cu121 # From CSM - ADJUST CUDA VERSION (cu118, cu121, etc.) AS NEEDED!
torchaudio==2.4.0 --extra-index-url https://download.pytorch.org/whl/cu121 # From CSM - ADJUST CUDA VERSION!

# Audio Handling
sounddevice
numpy
# pyaudio # Alternative/optional

# LLM Inference Engine (Choose ONE section)
# Option 1: vLLM
# vllm
vllm # Preferred engine for optimized inference

# Option 2: llama-cpp-python (Requires compilation with CUDA flags)
# llama-cpp-python

# Option 3: Hugging Face Transformers
# transformers>=4.38 # Or later version
transformers==4.49.0 # From CSM
accelerate
bitsandbytes # For 4/8-bit quantization
sentencepiece # Common tokenizer requirement
tokenizers==0.21.0 # From CSM

# VAD
silero-vad
# webrtcvad-wheels # Alternative

# Configuration
PyYAML # For loading config files
python-dotenv # For loading .env files

# Logging
# logging # Built-in

# Sesame AI Labs Tools (Install manually from cloned repos)
# whisperx (See repo for instructions)
# faster-whisper (See repo for instructions)
# moshi (See repo for instructions)
# gpt-fast (See repo for instructions)
# CSM Specific Dependencies
moshi==0.2.2
torchtune==0.4.0
torchao==0.9.0
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master

# Utilities
tqdm # Progress bars

# --- Hugging Face Hub ---
# huggingface-hub >= 0.20.0 # For model downloads & potentially CSM interaction
huggingface-hub==0.28.1 # From CSM

# --- STT ---
faster-whisper # Requires CTranslate2 backend, often bundled or installed separately if needed
# If build issues: pip install ctranslate2 -U # May need specific version

# --- LLM Engine ---
# vllm # Preferred engine for optimized inference # Already uncommented above
# sentencepiece # Often a dependency for tokenizers, install if needed by LLM/vLLM # Already included above

# --- RAG (Langchain + FAISS GPU) ---
langchain >= 0.1.0
langchain-community # Common loaders, stores etc.
langchain-huggingface # For HF embeddings if used
# Vector Store - FAISS for GPU acceleration
# faiss-gpu # Recommended for CUDA systems
faiss-gpu # Recommended for CUDA systems
# OR use faiss-cpu if GPU issues/preference
# faiss-cpu
# OR use ChromaDB
# chromadb
# Document Loaders (add more as needed)
pypdf # For PDFs
# python-docx # For .docx
# unstructured # For complex/varied document types
# Embeddings (using HF sentence-transformers)
sentence-transformers

# --- GUI (Example: CustomTkinter) ---
# customtkinter # Uncomment if using CustomTkinter for GUI
# Pillow # Often needed for images in GUI

# --- Sesame CSM Dependencies ---
# NOTE: Key dependencies like torch, transformers, etc., should align with
# requirements specified in external/csm/requirements.txt.
# Manually check external/csm/requirements.txt and add any MISSING dependencies here,
# or ensure versions above are compatible.
# Examples of potential CSM requirements (VERIFY IN external/csm/requirements.txt):
# transformers>=4.38 # Check CSM's required version # Added above with specific version
# accelerate # Included above
# bitsandbytes # If CSM uses quantization # Included above
# einops
# xformers # If used by CSM or its dependencies
# Note: CSM dependencies added near the top (moshi, torchtune, etc.) 